Clearer association between markers and cameras
As the camera identifier associated with a marker is simply bookkeeping for multicamera reconstruction, make it an optional attribute when handling tracks and reconstructed views. This means that the libmv API is still nice for users that don't need to associate images with cameras.

+++ extern/libmv/libmv-capi_stub.cc
@@ -85,8 +85,9 @@ struct libmv_Tracks *libmv_tracksNew(void)
 }
 
 void libmv_tracksInsert(struct libmv_Tracks * /*libmv_tracks*/,
-                        int /*view*/, int /*image*/, int /*track*/,
-                        double /*x*/, double /*y*/)
+                        int /*image*/, int /*track*/,
+                        double /*x*/, double /*y*/,
+                        int /*camera*/)
 {
 }
